Added direction class
Moved direction values into it's own class.
